+++ packages/headless/src/features/triggers/trigger-analytics-actions.ts
@@ -3,12 +3,28 @@ import {AnalyticsType, makeAnalyticsAction} from '../analytics/analytics-utils';
 /**
  * Log trigger query
  */
-export const logQueryTrigger = makeAnalyticsAction(
+export const logTriggerQuery = makeAnalyticsAction(
   'analytics/trigger/query',
   AnalyticsType.Search,
   (client, state) => {
     if (state.triggers?.query) {
       return client.logTriggerQuery();
+    }
+    return;
+  }
+);
+
+/**
+ * Log trigger notify
+ */
+export const logNotifyTrigger = makeAnalyticsAction(
+  'analytics/trigger/notify',
+  AnalyticsType.Search,
+  (client, state) => {
+    if (state.triggers?.notification) {
+      return client.logTriggerNotify({
+        notification: state.triggers.notification,
+      });
     }
     return;
   }