+++ packages/samples/headless-react/src/pages/SearchPage.tsx
@@ -52,6 +52,8 @@ import {QueryTrigger} from '../components/triggers/query-trigger.class';
 import {ExecuteTrigger} from '../components/triggers/execute-trigger.class';
 import {bindExecuteTrigger} from '../components/triggers/execute-trigger';
 import {QueryTrigger as QueryTriggerFn} from '../components/triggers/query-trigger.fn';
+import {NotifyTrigger} from '../components/triggers/notify-trigger.class';
+import {NotifyTrigger as NotifyTriggerFn} from '../components/triggers/notify-trigger.fn';
 import {SmartSnippet} from '../components/smart-snippet/smart-snippet.class';
 import {SmartSnippet as SmartSnippetFn} from '../components/smart-snippet/smart-snippet.fn';
 import {SmartSnippetQuestionsList} from '../components/smart-snippet-questions-list/smart-snippet-questions-list.class';
@@ -111,6 +113,8 @@ import {
   buildRedirectionTrigger,
   QueryTrigger as HeadlessQueryTrigger,
   buildQueryTrigger,
+  NotifyTrigger as HeadlessNotifyTrigger,
+  buildNotifyTrigger,
   SmartSnippet as HeadlessSmartSnippet,
   buildSmartSnippet,
   SmartSnippetQuestionsList as HeadlessSmartSnippetQuestionsList,
@@ -173,6 +177,7 @@ export class SearchPage extends Component {
   private readonly standaloneSearchBox: HeadlessStandaloneSearchBox;
   private readonly redirectionTrigger: HeadlessRedirectionTrigger;
   private readonly queryTrigger: HeadlessQueryTrigger;
+  private readonly notifyTrigger: HeadlessNotifyTrigger;
   private readonly smartSnippet: HeadlessSmartSnippet;
   private readonly smartSnippetQuestionsList: HeadlessSmartSnippetQuestionsList;
 
@@ -302,6 +307,8 @@ export class SearchPage extends Component {
     this.redirectionTrigger = buildRedirectionTrigger(this.engine);
 
     this.queryTrigger = buildQueryTrigger(this.engine);
+
+    this.notifyTrigger = buildNotifyTrigger(this.engine);
 
     this.smartSnippet = buildSmartSnippet(this.engine);
 
@@ -506,6 +513,10 @@ export class SearchPage extends Component {
           </Section>
           <Section title="execute-trigger">
             <ExecuteTrigger></ExecuteTrigger>
+          </Section>
+          <Section title="notify-trigger">
+            <NotifyTrigger></NotifyTrigger>
+            <NotifyTriggerFn controller={this.notifyTrigger}></NotifyTriggerFn>
           </Section>
         </AppContext.Provider>
       </div>